Extract utility functions to stateful classes
