+++ server/src/main/java/io/spine/server/entity/storage/Columns.java
@@ -23,24 +23,12 @@ package io.spine.server.entity.storage;
 import io.spine.annotation.Internal;
 import io.spine.server.entity.Entity;
 import io.spine.server.entity.storage.EntityColumn.MemoizedValue;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.beans.BeanInfo;
-import java.beans.IntrospectionException;
-import java.beans.Introspector;
-import java.beans.PropertyDescriptor;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.collect.Lists.newLinkedList;
-import static io.spine.server.entity.storage.ColumnRecords.getAnnotatedVersion;
-import static io.spine.util.Exceptions.newIllegalStateException;
 import static io.spine.validate.Validate.checkNotEmptyOrBlank;
 import static java.lang.String.format;
 
@@ -65,13 +53,6 @@ import static java.lang.String.format;
 @Internal
 public class Columns {
 
-    private static final String SPINE_PACKAGE = "io.spine.";
-    private static final String NON_PUBLIC_CLASS_WARNING =
-            "Passed entity class %s is not public. Storage fields won't be extracted.";
-    private static final String NON_PUBLIC_INTERNAL_CLASS_WARNING =
-            "Passed entity class %s is probably a Spine internal non-public entity. " +
-                    "Storage fields won't be extracted.";
-
     /**
      * Prevents instantiation of this utility class.
      */
@@ -81,17 +62,33 @@ public class Columns {
     /**
      * Ensures that the entity columns are valid for the specified entity class.
      *
-     * <p>This method tries to extract {@linkplain EntityColumn entity columns} from the given class,
-     * performing all checks along the way.
-     *
-     * <p>If extraction is performed without errors, the check is passed, if not - failed.
+     * <p>This method tries to obtain {@linkplain EntityColumn entity columns} for the given class,
+     * performing all checks along the way. If this process is performed without errors, the check is passed,
+     * if not - failed.
      *
      * @param entityClass the class to check entity columns
      */
     public static void checkColumnDefinitions(Class<? extends Entity> entityClass) {
         checkNotNull(entityClass);
 
-        obtainColumns(entityClass);
+        final ColumnReader columnReader = ColumnReader.forClass(entityClass);
+        columnReader.readColumns();
+    }
+
+    /**
+     * Retrieves {@linkplain EntityColumn columns} for the given {@code Entity} class.
+     *
+     * <p>Performs checks for entity column definitions correctness along the way.
+     *
+     * <p>If check for correctness fails, throws {@link IllegalStateException}.
+     *
+     * @param entityClass the class containing the {@link EntityColumn} definition
+     * @return a {@code Collection} of {@link EntityColumn} corresponded to entity class
+     * @throws IllegalStateException if entity column definitions are incorrect
+     */
+    static Collection<EntityColumn> getColumns(Class<? extends Entity> entityClass) {
+        final ColumnReader columnReader = ColumnReader.forClass(entityClass);
+        return columnReader.readColumns();
     }
 
     /**
@@ -108,7 +105,8 @@ public class Columns {
         checkNotNull(entityClass);
         checkNotEmptyOrBlank(columnName, "entity column name");
 
-        final Collection<EntityColumn> entityColumns = obtainColumns(entityClass);
+        final ColumnReader columnReader = ColumnReader.forClass(entityClass);
+        final Collection<EntityColumn> entityColumns = columnReader.readColumns();
         for (EntityColumn column : entityColumns) {
             if (column.getName()
                     .equals(columnName)) {
@@ -123,59 +121,25 @@ public class Columns {
     }
 
     /**
-     * Retrieves {@linkplain EntityColumn columns} for the given {@code Entity} class.
-     *
-     * <p>Performs checks for entity column definitions correctness along the way.
-     *
-     * <p>If check for correctness fails, throws {@link IllegalStateException}.
-     *
-     * @param entityClass the class containing the {@link EntityColumn} definition
-     * @return a {@link Collection} of {@link EntityColumn} corresponded to entity class
-     * @throws IllegalStateException if entity column definitions are incorrect
-     */
-    static Collection<EntityColumn> obtainColumns(Class<? extends Entity> entityClass) {
-        checkNotNull(entityClass);
-
-        final BeanInfo entityDescriptor;
-        try {
-            entityDescriptor = Introspector.getBeanInfo(entityClass);
-        } catch (IntrospectionException e) {
-            throw new IllegalStateException(e);
-        }
-
-        final Collection<EntityColumn> entityColumns = newLinkedList();
-        for (PropertyDescriptor property : entityDescriptor.getPropertyDescriptors()) {
-            final Method getter = property.getReadMethod();
-            final boolean isEntityColumn = getAnnotatedVersion(getter).isPresent();
-            if (isEntityColumn) {
-                final EntityColumn column = EntityColumn.from(getter);
-                entityColumns.add(column);
-            }
-        }
-        checkRepeatedColumnNames(entityColumns, entityClass);
-
-        return entityColumns;
-    }
-
-    /**
      * Extracts the {@linkplain EntityColumn column} values for the given {@linkplain Entity}.
      *
      * <p>Retrieves {@linkplain EntityColumn columns} for the given {@code Entity} class, then generates
      * {@linkplain MemoizedValue memoized values} from them.
      *
-     * <p>This method will return {@linkplain Collections#emptyMap() empty map} for non-public
-     * {@link Entity} classes.
+     * <p>This method will return {@linkplain Collections#emptyMap() empty map} for {@link Entity} classes
+     * that are non-public or cannot be subjected to column extraction for some other reason.
      *
      * @param entity an {@link Entity} to get the {@linkplain EntityColumn column} values from
      * @param <E>    the type of the {@link Entity}
-     * @return a {@link Map} of the column {@linkplain EntityColumn#getStoredName()
+     * @return a {@code Map} of the column {@linkplain EntityColumn#getStoredName()
      *         names for storing} to their {@linkplain MemoizedValue memoized values}
      * @see MemoizedValue
      */
     static <E extends Entity<?, ?>> Map<String, EntityColumn.MemoizedValue> extractColumnValues(E entity) {
         checkNotNull(entity);
 
-        final Collection<EntityColumn> entityColumns = obtainColumns(entity.getClass());
+        final ColumnReader columnReader = ColumnReader.forClass(entity.getClass());
+        final Collection<EntityColumn> entityColumns = columnReader.readColumns();
         return extractColumnValues(entity, entityColumns);
     }
 
@@ -184,7 +148,7 @@ public class Columns {
      *
      * <p>Uses given {@linkplain EntityColumn entity columns} for the value extraction.
      *
-     * <p>This way the process of {@linkplain Columns#obtainColumns(Class) obtaining columns} from
+     * <p>This way the process of {@linkplain ColumnReader#readColumns() obtaining columns} from
      * the given {@link Entity} class can be skipped.
      *
      * <p>This method will return {@linkplain Collections#emptyMap() empty map} for {@link Entity} classes
@@ -193,109 +157,16 @@ public class Columns {
      * @param entity        an {@link Entity} to get the {@linkplain EntityColumn column} values from
      * @param entityColumns {@linkplain EntityColumn entity columns} which values should be extracted
      * @param <E>           the type of the {@link Entity}
-     * @return a {@link Map} of the column {@linkplain EntityColumn#getStoredName()
-     *         names for storing} to their {@linkplain MemoizedValue memoized values}
-     * @see MemoizedValue
-     */
-    static <E extends Entity<?, ?>> Map<String, MemoizedValue>
-    extractColumnValues(E entity, Collection<EntityColumn> entityColumns) {
-        checkNotNull(entityColumns);
-        checkNotNull(entity);
-
-        final Class<? extends Entity> entityClass = entity.getClass();
-        if (!canExtractColumns(entityClass)) {
-            return Collections.emptyMap();
-        }
-
-        final Map<String, MemoizedValue> values = readValues(entityColumns, entity);
-        return values;
-    }
-
-    /**
-     * Checks if the given {@link Entity entity class} is public.
-     *
-     * <p>Outputs a message to the log if the class is non-public.
-     *
-     * @param entityClass {@link Entity entity class} to check
-     * @return {@code true} if class is public and {@code false} otherwise
-     */
-    private static boolean canExtractColumns(Class<? extends Entity> entityClass) {
-        checkNotNull(entityClass);
-
-        final int modifiers = entityClass.getModifiers();
-        if (!Modifier.isPublic(modifiers)) {
-            logNonPublicClass(entityClass);
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Extracts {@linkplain MemoizedValue memoized values} for the given {@linkplain EntityColumn columns}
-     * of the given {@link Entity}.
-     *
-     * @param columns {@link Collection collection} of columns to extract values from
-     * @param entity  {@link Entity} from which to extract the values
      * @return a {@code Map} of the column {@linkplain EntityColumn#getStoredName()
-     *         names for storing} to their {@linkplain MemoizedValue memoized values}
-     * @see MemoizedValue
+     *         names for storing} to their {@linkplain EntityColumn.MemoizedValue memoized values}
+     * @see EntityColumn.MemoizedValue
      */
-    private static <E extends Entity<?, ?>> Map<String, MemoizedValue> readValues(Collection<EntityColumn> columns,
-                                                                                  E entity) {
-        final Map<String, MemoizedValue> values = new HashMap<>(columns.size());
-        for (EntityColumn column : columns) {
-            final String name = column.getStoredName();
-            final MemoizedValue value = column.memoizeFor(entity);
-            values.put(name, value);
-        }
-        return values;
-    }
+    static <E extends Entity<?, ?>> Map<String, EntityColumn.MemoizedValue>
+    extractColumnValues(E entity, Collection<EntityColumn> entityColumns) {
+        checkNotNull(entity);
+        checkNotNull(entityColumns);
 
-    /**
-     * Ensures that the specified columns have no repeated names.
-     *
-     * <p>If the check fails, throws {@link IllegalStateException}.
-     *
-     * @param columns     the columns to check
-     * @param entityClass the entity class for the columns
-     * @throws IllegalStateException if columns contain repeated names
-     */
-    private static void checkRepeatedColumnNames(Iterable<EntityColumn> columns,
-                                                 Class<? extends Entity> entityClass) {
-        final Collection<String> checkedNames = newLinkedList();
-        for (EntityColumn column : columns) {
-            final String columnName = column.getStoredName();
-            if (checkedNames.contains(columnName)) {
-                throw newIllegalStateException(
-                        "The entity `%s` has columns with the same name for storing `%s`.",
-                        entityClass.getName(),
-                        columnName);
-            }
-            checkedNames.add(columnName);
-        }
-    }
-
-    /**
-     * Writes the non-public {@code Entity} class warning into the log unless the passed class
-     * represents one of the Spine internal {@link Entity} implementations.
-     */
-    private static void logNonPublicClass(Class<? extends Entity> cls) {
-        final String className = cls.getCanonicalName();
-        final boolean internal = className.startsWith(SPINE_PACKAGE);
-        if (internal) {
-            log().trace(format(NON_PUBLIC_INTERNAL_CLASS_WARNING, className));
-        } else {
-            log().warn(format(NON_PUBLIC_CLASS_WARNING, className));
-        }
-    }
-
-    private static Logger log() {
-        return LogSingleton.INSTANCE.value;
-    }
-
-    private enum LogSingleton {
-        INSTANCE;
-        @SuppressWarnings("NonSerializableFieldInSerializableClass")
-        private final Logger value = LoggerFactory.getLogger(Columns.class);
+        final ColumnValueExtractor columnValueExtractor = ColumnValueExtractor.create(entity, entityColumns);
+        return columnValueExtractor.extractColumnValues();
     }
 }
